Split Editors | Move inEditorResultsActions to ResultsSplitEditor interface
Promotes inEditorResultsActions() from a protected SplitEditorBase hook to a
documented public interface method on ResultsSplitEditor with a default empty-list
implementation. Language-specific editors can now override it directly on their
interface or concrete class without inheriting from SplitEditorBase. KDoc explains
placement (left side of bar), intended use (export, copy, filter actions), and
default behaviour. Skills updated to reflect DataKey ownership on domain interface
companion objects.
