--- modules/shared/ui/src/sap/commerce/toolset/ui/editor/HideInEditorResultsAction.kt
@@ -1,42 +1,0 @@
-/*
- * This file is part of "SAP Commerce Developers Toolset" plugin for IntelliJ IDEA.
- * Copyright (C) 2019-2026 EPAM Systems <hybrisideaplugin@epam.com> and contributors
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Lesser General Public License as
- * published by the Free Software Foundation, either version 3 of the
- * License, or (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
- * See the GNU Lesser General Public License for more details.
- *
- * You should have received a copy of the GNU Lesser General Public License
- * along with this program. If not, see <http://www.gnu.org/licenses/>.
- */
-
-package sap.commerce.toolset.ui.editor
-
-import com.intellij.icons.AllIcons
-import com.intellij.openapi.actionSystem.ActionUpdateThread
-import com.intellij.openapi.actionSystem.AnActionEvent
-import com.intellij.openapi.project.DumbAwareAction
-
-class HideInEditorResultsAction : DumbAwareAction(
-    "Hide",
-    "Hide in-editor results panel",
-    AllIcons.Actions.Close
-) {
-
-    override fun getActionUpdateThread() = ActionUpdateThread.BGT
-
-    override fun update(e: AnActionEvent) {
-        e.presentation.isEnabledAndVisible = e.getData(SplitEditorDataKeys.SPLIT_EDITOR) != null
-    }
-
-    override fun actionPerformed(e: AnActionEvent) {
-        val editor = e.getData(SplitEditorDataKeys.SPLIT_EDITOR) ?: return
-        editor.inEditorResults = !editor.inEditorResults
-    }
-}
